+++ deptry/issue_finders/transitive_dependencies_finder.py
@@ -2,21 +2,19 @@ import logging
 from typing import List
 
 from deptry.dependency import Dependency
+from deptry.issue_finders.issue_finder import IssueFinder
 from deptry.module import Module
 
 
-class TransitiveDependenciesFinder:
+class TransitiveDependenciesFinder(IssueFinder):
     """
     Given a list of imported modules and a list of project dependencies, determine which ones are transitive.
     """
 
     def __init__(
-        self, imported_modules: List[Module], dependencies: List[Dependency], ignore_transitive: List[str] = []
+        self, imported_modules: List[Module], dependencies: List[Dependency], list_to_ignore: List[str] = []
     ) -> None:
-        self.imported_modules = imported_modules
-        self.imported_modules = self._filter_out_standard_library_from(self.imported_modules)
-        self.dependencies = dependencies
-        self.ignore_transitive = ignore_transitive
+        super().__init__(imported_modules, dependencies, list_to_ignore)
 
     def find(self) -> List[str]:
         logging.debug("\nScanning for transitive dependencies...")
@@ -26,31 +24,17 @@ class TransitiveDependenciesFinder:
     def _get_transitive_dependencies(self):
         transitive_dependencies = []
         for module in self.imported_modules:
+            logging.info(f"Scanning module {module.name}...")
             if (
                 module.package is not None
                 and not self._module_in_any_top_level(module)
                 and not self._module_in_dependencies(module)
-                and not module.local_module
+                and not module.is_local_module()
             ):
-                if module.name in self.ignore_transitive:
+                if module.name in self.list_to_ignore:
                     logging.debug(f"Module '{module.package}' found to be a transitive dependency, but ignoring.")
                 else:
                     transitive_dependencies.append(module.package)
                     logging.debug(f"Dependency '{module.package}' marked as a transitive dependency.")
 
         return transitive_dependencies
-
-    def _module_in_any_top_level(self, module: Module) -> bool:
-        for dependency in self.dependencies:
-            if dependency.top_levels and module.name in dependency.top_levels:
-                return True
-        return False
-
-    def _module_in_dependencies(self, module: Module) -> bool:
-        for dependency in self.dependencies:
-            if module.package == dependency.name:
-                return True
-        return False
-
-    def _filter_out_standard_library_from(self, imported_modules: List[Module]):
-        return [module for module in imported_modules if not module.standard_library]
